refactor(presentation): 보이스노트 포맷팅 로직을 Core 익스텐션으로 분리 및 리팩터링
